Don't execute bin2img if image was created already with 'newt target label'.

+++ hw/bsp/nrf51dk-16kbram/nrf51dk-16kbram_download.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -49,11 +49,15 @@ if [ $IS_BOOTLOADER -eq 1 ]; then
     FILE_NAME=$BASENAME.elf.bin
 else
     FLASH_OFFSET=0x7000
-    FILE_NAME=$BASENAME.elf.img
-    echo "Version is >" $VER "<"
-    $BIN2IMG $BASENAME.elf.bin $FILE_NAME $VER
-    if [ "$?" -ne 0 ]; then
-	exit 1
+    ELF_FILE=$BASENAME.elf
+    BIN_FILE=$BASENAME.elf.bin
+    FILE_NAME=$BASENAME.img
+    if [ \! -f $FILE_NAME -o $FILE_NAME -ot $ELF_FILE ]; then
+        echo "Version is >" $VER "<"
+        $BIN2IMG $BASENAME.elf.bin $FILE_NAME $VER
+        if [ "$?" -ne 0 ]; then
+	    exit 1
+	fi
     fi
 fi
